Added: prevent muting staff

+++ commands/Staff/mute.js
@@ -6,14 +6,18 @@ module.exports.run = async (bot, message, args) => {
     if (!user){
         userArray = await message.channel.guild.fetchMembers({userIDs:[args[0]]})
         user = userArray[0]
+    }else{
+        userArray = await message.channel.guild.fetchMembers({userIDs:[user.id]})
+        user = userArray[0]
     }
+    if (user.roles.includes('704921933073875014')) return message.channel.createMessage("This member can not be muted")
     eventChannel = bot.getChannel("779081002311352370")
     eventChannel.editPermission(user.id, 0,2048,"member","Event mute")
     eventChannel.createMessage(`User: ${user.username}#${user.discriminator} has been muted`)
     setTimeout(() => {
         eventChannel.deletePermission(user.id, "Event unmute")
         eventChannel.createMessage(`User: ${user.username}#${user.discriminator} has been unmuted`)
-    }, ms("20s"));
+    }, ms("30m"));
     
 }
